Columbus/backend/requirements.txt
asgiref==3.4.1
awscli==1.22.23
boto3==1.20.23
botocore==1.23.23
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.9
django-cors-headers==3.10.0
django-env==0.2.1
django-environ==0.8.1
django-rest-swagger==2.1.2
djangorestframework==3.12.4
docutils==0.15.2
environ==1.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
jmespath==0.10.0
MarkupSafe==2.0.1
openapi-codec==1.3.2
psycopg2-binary==2.9.1
pyasn1==0.4.8
python-dateutil==2.8.2
pytz==2021.3
PyYAML==5.4.1
pyyml==0.0.2
requests==2.26.0
rsa==4.7.2
s3transfer==0.5.0
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
